fix: version 24 error

+++ zrlog-common/src/main/resources/conf/update-sql/24.sql
@@ -1,7 +1,7 @@
 ALTER TABLE log ADD COLUMN extensions longtext DEFAULT NULL;
 CREATE TABLE IF NOT EXISTS `log_extension_index`
 (
-    `id`              bigint(20)   NOT NULL AUTO_INCREMENT,
+    `id`              int(11)   NOT NULL AUTO_INCREMENT,
     `log_id`          int(11)      NOT NULL,
     `namespace`       varchar(64)  NOT NULL,
     `extension_path`  varchar(191) NOT NULL,
